requirements.txt
Django==1.9.1
django-appconf==1.0.1
django-compressor==2.0
django-crispy-forms==1.6.0
django-filter==0.12.0
djangorestframework==3.3.2
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
wheel==0.24.0
